Solve Retina display issues
Initially x1,y1,x2,y2 coordinates in pixel where multiplied by 2 on retina displays but when grabbing you divide by 2 again. This does not solve the issue which is related only to grab where only the box captured must be scaled. Hence the proposed change solve the issue and you can verify with img.show() to check the part of the screen captured